pyproject.toml
[project]
name = "doubt-server-beta"
version = "0.1.0"
description = "Socket.IO Chat Server with FastAPI"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-socketio>=5.11.0",
    "httpx>=0.28.1",
]
requires-python = ">=3.8"

[project.scripts]
doubt-server = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
